- added solutionWizard
