+++ src/TemplateBuilder/TemplateBuilder.csproj
@@ -55,6 +55,7 @@
     <Compile Include="ChildWizard.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RootWizard.cs" />
+    <Compile Include="SolutionWizard.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
